refactor: change module display name to "UpCloud Server"
UpCloud is spelled with a capital "U", and we do not currently use the
term "VPS" with our products.

The previous version of the UpCloud WHMCS module used "UpCloud VM" which
would likely be ok as well, but better to differentiate in case both old
and this version of the module are both installed.